Add utility tests for email and password handling, including hashing and masking
