+++ tests/test_routes/conftest.py
@@ -1,0 +1,9 @@
+import pytest
+from flask import Flask
+
+
+@pytest.fixture
+def client(app: Flask):
+    """Create a test client for the Flask application"""
+    with app.test_client() as client:
+        yield client
