+++ tests/test_routes/test_authentication/test_register.py
@@ -1,38 +1,4 @@
-import pytest
-from flask import Flask
 from flask.testing import FlaskClient
-
-from routes.authentication import authentication_blueprint
-
-
-@pytest.fixture
-def app():
-    """Create and configure a new Flask application instance for testing"""
-    app = Flask(__name__)
-    app.register_blueprint(authentication_blueprint)
-    return app
-
-
-@pytest.fixture
-def client(app: Flask):
-    """Create a test client for the Flask application"""
-    with app.test_client() as client:
-        yield client
-
-
-@pytest.fixture
-def sample_email():
-    return "sample.email@example.com"
-
-
-@pytest.fixture
-def sample_password():
-    return "SecurePass123!"
-
-
-@pytest.fixture
-def sample_username():
-    return "sampleuser"
 
 
 def test_missing_email(client: FlaskClient, sample_username, sample_password):
