+++ tests/test_utility/test_email/conftest.py
@@ -1,0 +1,6 @@
+import pytest
+
+
+@pytest.fixture
+def sample_email():
+    return "sample.email@example.com"
