+++ tests/test_utility/test_password/conftest.py
@@ -1,0 +1,6 @@
+import pytest
+
+
+@pytest.fixture
+def sample_password():
+    return "SecurePass123!"
